Use for appveyor
